Insert null checks before base/this call if arg is used
If the constructor arguments are being used in the base/this call, then
insert null checks for those arguments before that call.

Note: IMO this should only happen for args that are dereferenced in the
base/this call, not for arguments that are merely forwarded. Ideally,
forwarded arguments that aren't used in the caller constructor should be
validated by the callee only, but this commit is not implementing that
logic - all arguments are validated everywhere at present.